feat: user reviews view w/o caching + other stuff
