+++ FoodBookApp/Repositories/Review/ReviewRepositoryImpl.swift
@@ -23,4 +23,8 @@ class ReviewRepositoryImpl: ReviewRepository {
     func uploadReviewReport(reviewId: String, reason: String) async throws {
         return try await ReviewRepositoryImpl.reviewSA.uploadReviewReport(reviewId: reviewId, reason: reason)
     }
+    
+    func getUserReviews() async throws -> [Review] {
+        return try await ReviewRepositoryImpl.reviewSA.getUserReviews()
+    }
 }
