+++ FoodBookApp/ServiceAdapter/ReviewSA.swift
@@ -13,4 +13,5 @@ protocol ReviewSA {
     func createReview(review: Review) async throws -> String
     func uploadPhoto(image: UIImage) async throws -> String
     func uploadReviewReport(reviewId: String, reason: String) async throws
+    func getUserReviews() async throws -> [Review]
 }
